#32 clean the code and create EmailSendingService
