+++ src/main/java/org/matrix/izumbankapp/dao/repository/DepositRepository.java
@@ -11,11 +11,6 @@ import java.util.List;
 public interface DepositRepository extends JpaRepository<DepositEntity, Long> {
     DepositEntity findByAccountId(Long id);
 
-//    @EntityGraph("Deposit.withAccount")
-//    List<DepositEntity> findDepositAccountsCreatedOnDate(@Param("dayOfMonth") int dayOfMonth);
-//    @Query("SELECT d FROM DepositEntity d LEFT JOIN FETCH d.account WHERE FUNCTION('DAY', d.account.createdAt) = :dayOfMonth")
-//    List<DepositEntity> findDepositsCreatedOnDate(@Param("dayOfMonth") int dayOfMonth);
-
     @Query("SELECT d FROM DepositEntity d LEFT JOIN FETCH d.account WHERE EXTRACT(DAY FROM d.account.createdAt) = :dayOfMonth")
     List<DepositEntity> findDepositsCreatedOnDate(@Param("dayOfMonth") int dayOfMonth);
