+++ src/main/java/org/matrix/izumbankapp/service/DepositService.java
@@ -7,14 +7,8 @@ import java.util.List;
 
 @Service
 public interface DepositService {
-    List<DepositResponse> getAllDeposits();
-
-    void saveDeposits(List<DepositResponse> depositResponses);
 
     void saveDeposit(DepositResponse depositResponse);
-
-    DepositResponse getDepositByAccountId(Long id);
-
     List<DepositResponse> getDepositAccountsCreatedOnDate(int dayOfMonth);
 
 }
